Pull in images base URL from bootstrapped CONFIGS (redo) (#42)
# Changelog

Combines #40 and #41 with some changes.

Set us up to stop hardcoding `/static/images/` as the base of all image URLs; expose a place to bootstrap this value into a `CONFIGS` object, which the new `configs` module consumes, allowing any component to import it as needed. Use this value react-style in `IframeCarousel` for its buttons, and in generating the `Save` icon.