Update interface of models to match between boostrapped JSON and API responses (#47)
Currently, in order to render video and audio carousels, we bootstrap each iframe's `src` property into the DOM in a JSON array and extract them in react. This allows us to avoid many pitfalls and security holes that could result from parsing user-provided values on the client too readily. (The situation is similar for photos.)

Also currently, in order to modify existing videos associated with a user's account, the client makes an API call and gets an object that contains the video's `code` attribute, but no parsed-out `src` value.

We want to share `video` data between the carousel and the form so that the carousel can be updated client-side after modifications are successfully made to the `video` data via the form. It will be much cleaner to do this sharing if the structure for each representation of the `video` data matches.

Critically, the carousel specifically needs access to the parsed-out `src` value that will be given to the iframe, while the form needs each video's full `code`. Exposing both `src` and `code` in one interface is a little duplicative, since the `src` value is encompassed in the `code` string, but it solves this problem well, without treating user-provided data as safe, and without flakiness associated with e.g. un-unit-tested regex running client-side for extracting the `src` value.

Very open to other solutions; this PR is to demonstrate the intention as much as it is an available solution.