Genericize video edit form (#55)
This is a redo of #50.

# Changelog

For [adding the audio edit form](https://trello.com/c/H3RcjdBu/8-artist-profile-edit-audio).

This PR genericizes the video edit form so that the audio edit form can be built on its infrastructure. This PR also separates out some components into their own files for legibility.

Note that, to be reused for the photo edit form, this code will have to be refactored again slightly; this refactor is only a first pass. As such, a few places where the code could hypothetically be DRYer without consideration of the photo edit form are left as-is.

## Changes
* Refactor the video edit form with more reusable code.
* Stop the modal from popping when the user switches between tabs of different heights.
* Make 'focus' color for tabs less obnoxious.

## New
* Explicitly define DragHandleMoveButton, DraggableCodeInputs, Header, and HelpSection components in their own files for legibility.
* Create `unicodeCharacters` file for referencing unicode characters by code point. (In response to [this comment](https://github.com/OpusApps/booking/pull/50#discussion_r213399069), because using html entities in JS strings without double-escaping [is not as simple as one might hope](https://shripadk.github.io/react/docs/jsx-gotchas.html). )

https://github.com/OpusApps/booking/pull/55