Force first photo = cover photo (#91)
For [this trello card](https://trello.com/c/ilUw74ED/12-artist-profile-edit-photos)

This PR triggers a new request the first time after a user uploads a new photo when they have no existing cover photo. The most obvious use case is when the user is adding photos for the first time. It will also happen when the user has deleted a cover photo and not yet assigned a new one, which will come up less often.

This could be done differently by updating the cover photo on the server during the same request, and it could be better UX that way (because of fewer round-trips e.g.), but would require changes on both the client and the server.

https://github.com/OpusApps/booking/pull/91