Update relative filepath for hamburger icon (#135)
At some point, the relative filepath for the hamburger icon given in a sass file became incorrect. The filepath on master now resolves to `root/static/scss/images/ico-menu.svg`, using a path relative to `root/static/scss/home`, but the icon is located at `root/static/images/home/ico-menu.svg`. Not immediately obvious what got swapped when, but we may want to reconsider using relative filepaths in our sass files (out of scope for this PR though).

This PR corrects the filepath for the hamburger icon so it is once again visible on the home screen on mobile devices.

For [this trello card](https://trello.com/c/i8NImk4Y/234-log-in-sign-up-on-mobile-is-not-apparent-unless-you-scroll-to-links-at-the-bottom-of-the-page)