Fix artist card profile image sideways orientation (#136)
This PR adds a function to get an auto-oriented image from cloudinary without any other image transformations, then uses this auto-oriented image instead of the default in the artist card profile photo. (All other image transformations seem to auto-orient the image without that transformation specified, so it is not necessary for e.g. thumbnails to include the auto-orient flag in their URLs.)

We may want a solution that is less manual and fragile, like always adding the auto-orient flag when a cloudinary image URL enters the redux store.

For half of [this trello card](https://trello.com/c/gQh7EMw1/241-uploading-a-photo-during-onboarding-flow-and-on-edit-icon-in-the-artist-card-will-results-in-the-photos-to-upload-sideways)