Unrequire avatar and improve empty states (#143)
This makes the avatar optional during onboarding and makes an empty avatar clickable on the profile to open the edit form (but only if its the profile owner viewing their own profile). This also addresses making the _cover photo_ empty state more prominent. 

Task is tracked in trello here: https://trello.com/c/IqeA9uAv/262-avatar-onboarding-improvements-not-required-in-signup-better-empty-state-on-profile