Master -> Production (#147)
* Production (#140)

* Sync master -> Production (#139)

* Use validator to check URLs (#138)

* fix: Use validator to check URLs

* fix: Re-order as per eslint

* chg: Turn off SSL redirect in favor of ALB

* fix: Implement SSL redirect elsewhere for now

* A few quick and easy fixes (#141)

* Misc Fixes (#145)

* Unrequire avatar and improve empty states (#143)

This makes the avatar optional during onboarding and makes an empty avatar clickable on the profile to open the edit form (but only if its the profile owner viewing their own profile). This also addresses making the _cover photo_ empty state more prominent. 

Task is tracked in trello here: https://trello.com/c/IqeA9uAv/262-avatar-onboarding-improvements-not-required-in-signup-better-empty-state-on-profile